added the intersect function, split it into two different functions

+++ structs.h
@@ -2,8 +2,9 @@
 #define __STRUCTS_H__
 
 typedef struct{
-	float x, y, z;
+ double x, y, z;
 }Vector3D;
+
 
 typedef struct{
 	float x, y, z, w;
